+++ web/runs/runs.html
@@ -17,7 +17,6 @@
 		<script src="/components/sidebar.js"></script>
 		<script src="/components/footer.js"></script>
 		<script src="/components/loading-spinner.js"></script>
-		<script src="/components/progress-bar.js"></script>
 		<script src="/components/confirm-dialog.js"></script>
 		<script src="/components/submit-form-button.js"></script>
 		<script src="/components/run-form.js"></script>
@@ -25,7 +24,7 @@
 		<script src="/scripts/main.js"></script>
 		<script src="/scripts/api.js"></script>
 		<script src="/scripts/runs.js"></script>
-		<script src="/scripts/run-utilities.js"></script>
+		<script src="/scripts/runs-shared.js"></script>
 
 		<script src="/scripts/lib/dayjs.min.js"></script>
 		<script src="/scripts/lib/dayjs.timezone.js"></script>
