+++ web/scripts/runs.js
@@ -3,20 +3,6 @@ let shoes = [];
 
 window.addEventListener('load', function () {
 	loadRuns();
-	loadRunningShoes();
-
-	document.querySelector('#add-new-run')?.addEventListener('click', function () {
-		openAddRunForm();
-	});
-
-	document.querySelector('#form-dialog-cancel').addEventListener('click', function () {
-		closeShoeForm();
-	});
-
-	document.querySelector('form[name="run-form"]')?.addEventListener('submit', async function (e) {
-		e.preventDefault();
-		await saveRun();
-	});
 
 	document.querySelector('dialog.confirm-dialog button.confirm-dialog-no').addEventListener('click', function () {
 		cancelOutOfConfirmDialog('#delete-run-id');
@@ -26,33 +12,6 @@ window.addEventListener('load', function () {
 		deleteRun();
 	});
 });
-
-async function loadRunningShoes() {
-	const [data, error] = await Api.Get('shoe');
-
-	if (error) {
-		showPageError(error);
-		return;
-	}
-
-	if (data.length === 0) {
-		shoes = [];
-		return;
-	}
-
-	shoes = data.map((d) => {
-		if (d.isRetired) {
-			return {
-				...d,
-				name: `${d.name} (Retired)`,
-			};
-		}
-
-		return d;
-	});
-
-	document.getElementById('shoe-id').appendChild(buildSelectList(shoes, 'shoeId', 'name'));
-}
 
 async function loadRuns() {
 	const [data, error] = await Api.Get('run');
@@ -96,7 +55,7 @@ function loadTable() {
 		tr.querySelector('.date-col').textContent = dateRun.format('M/D/YYYY');
 		tr.querySelector('.temperature-col').textContent = run.temperature ? `${run.temperature}ºF` : '';
 		tr.querySelector('.time-col').textContent = run.runTime;
-		tr.querySelector('.distance-col').textContent = run.distance;
+		tr.querySelector('.distance-col').textContent = run.distance.toFixed(2);
 		tr.querySelector('.pace-col').textContent = `${run.pace}/mi.`;
 		tr.querySelector('.elevation-col').textContent = run.elevation ? run.elevation : '';
 		tr.querySelector('.hr-col').textContent = run.heartRate ? run.heartRate : '';
@@ -154,35 +113,6 @@ function buildMonthlySummaryRow(rows) {
 	return tr;
 }
 
-function setProcessing(isProcessing) {
-	const btn = document.querySelector('submit-form-button.save-run-submit');
-
-	if (isProcessing) {
-		btn.setAttribute('processing', 'true');
-		btn.setAttribute('buttontext', 'Saving');
-		document.querySelectorAll('form[name="run-form"] button').forEach((b) => b.setAttribute('disabled', 'true'));
-	
-	} else {
-		btn.setAttribute('processing', 'false');
-		btn.setAttribute('buttontext', 'Save');
-		document.querySelectorAll('form[name="run-form"] button').forEach((b) => b.removeAttribute('disabled'));
-	}
-}
-
-function openAddRunForm() {
-	document.querySelector('#run-form-dialog h3').textContent = 'Add Run';
-	document.querySelector('dialog#run-form-dialog input#runId').value = '0';
-	document.querySelector('dialog#run-form-dialog input#date-run').value = dayjs().format('YYYY-MM-DD');
-
-	const activeShoes = shoes.filter((s) => !s.isRetired);
-
-	if (activeShoes.length === 1) {
-		document.querySelector('select#shoe-id').value = activeShoes[0].shoeId;
-	}
-
-	document.querySelector('dialog#run-form-dialog').showModal();
-}
-
 function editRun(run) {
 	if (run) {
 		document.querySelector('#run-form-dialog h3').textContent = 'Edit Run';
@@ -199,64 +129,6 @@ function editRun(run) {
 
 		document.querySelector('dialog#run-form-dialog').showModal();
 	}
-}
-
-function closeShoeForm() {
-	document.querySelector('form[name="run-form"]').reset();
-	document.querySelector('dialog#run-form-dialog').close();
-	setProcessing(false);
-}
-
-function buildRunData() {
-	return {
-		runId: parseInt(document.querySelector('input#runId').value),
-		dateRan: document.querySelector('input#date-run').value,
-		distance: parseFloat(document.querySelector('input#distance').value),
-		hours: parseInt(document.querySelector('input#hours').value),
-		minutes: parseInt(document.querySelector('input#minutes').value),
-		seconds: parseInt(document.querySelector('input#seconds').value),
-		temperature: parseInt(document.querySelector('input#temperature').value),
-		heartRate: parseInt(document.querySelector('input#heart-rate').value),
-		elevation: parseInt(document.querySelector('input#elevation').value),
-		shoeId: parseInt(document.querySelector('select#shoe-id').value),
-	}
-}
-
-async function addRun(values) {
-	const [, error] = await Api.Post('run', {
-		data: values,
-	});
-
-	return error;
-}
-
-async function updateRun(values) {
-	const [, error] = await Api.Put(`run/${values.runId}`, {
-		data: values,
-	});
-
-	return error;
-}
-
-async function saveRun() {
-	hideModalError('modal-error');
-	setProcessing(true);
-
-	const values = buildRunData();
-
-	const error = values.runId > 0
-		? await updateRun(values)
-		: await addRun(values);
-
-	if (error) {
-		showModalError(error, 'modal-error');
-		setProcessing(false);
-		return;
-	}
-
-	await loadRuns();
-
-	closeShoeForm();
 }
 
 function openDeleteConfirmation(run) {
