+++ web/styles/main.css
@@ -835,6 +835,17 @@ ul.tag-list li {
 .bolded {
 	font-weight: 700;
 }
+
+.flex-header {
+	display: flex;
+	align-items: center;
+	justify-content: space-between;
+	margin-bottom: 24px;
+}
+
+.flex-header h1 {
+	margin-bottom:0;
+}
 /* END UTILITY CLASSES */
 
 /* CHARTS */
