+++ web/styles/runs.css
@@ -34,3 +34,9 @@ table {
 		width: 100%;
 	}
 }
+
+.total-time {
+	display:flex;
+	flex-direction: column;
+	row-gap: 4px;
+}
